# public/swagger.yaml
openapi: 3.0.0
info:
  title: Boutique Management API
  version: 1.0.0
  description: API documentation for the Boutique Management application
servers:
  - url: http://localhost:8000/api  
security:
  - BearerAuth: []
paths:
...
  /v1/role:
    post:
      summary: Create a new role
      tags:
        - Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreRoleRequest'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users:
    get:
      summary: Get a list of users
      tags:
        - Users
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [ADMIN, BOUTIQUIER, CLIENT]
          description: Filter users by role
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/{id}:
    get:
      summary: Get a user by ID
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to retrieve
      responses:
        '200':
          description: A single user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: Invalid request
...
    Role:
      type: object
      properties:
        id:
          type: string
          example: 123
        name:
          type: string
          example: Admin
    User:
      type: object
      properties:
        id:
          type: string
          example: 123
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: user@example.com
        role:
          type: string
          example: Admin
...
    StoreRoleRequest:
      type: object
      properties:
        name:
          type: string
          example: Admin
    StoreUserRequest:
      type: object
      properties:
        nom:
          type: string
          example: Doe
        prenom:
          type: string
          example: John
        password:
          type: string
          example: Pourquoipas34$
        password_confirmation:
          type: string
          example: Pourquoipas34$
        role_id:
          type: integer
          example: 1
        etat:
          type: boolean
          example: true
        login:
          type: string
          example: bakemono
...